refactored some code to allow for main app testing
