+++ www/app.js
@@ -2,11 +2,11 @@ const { App } = require("@slack/bolt");
 const dayjs = require('dayjs');
 const utc = require('dayjs/plugin/utc');
 const timezone = require('dayjs/plugin/timezone');
-const { find_today } = require("wasm-build");
 require("dotenv").config();
-const navigate = require('./app/navigate');
-const { schedule, validateWebScrapingTime, calculateInterval, sleep } = require('./app/scheduler');
+const { runTimeOffEvents } = require('./app/helper');
 
+dayjs.extend(utc);
+dayjs.extend(timezone);
 // Initializes your app with your bot token and signing secret
 const app = new App({
   token: process.env.SLACK_BOT_TOKEN,
@@ -42,37 +42,6 @@ app.command('/time_off/help', async ({ command, ack, respond})=>{
   }
 });
 
-async function runTimeOffEvents(startTime) {
-  // calculate timeout ms
-  const sleepTime = calculateInterval(startTime);
-  console.log(sleepTime);
-  await sleep(sleepTime);
-  const timeInterval = 8.64e7;
-  setInterval(async () => {
-    if (validateWebScrapingTime(date)) {
-      try {
-        console.time('nav');
-        const html = await navigate('https://hr.humi.ca/login');
-        console.timeEnd('nav');
-        const timeOff = find_today(html);
-        // const timeOff = ['test'];
-        if (timeOff.length > 0 ) {
-          const post_at = schedule(date); 
-          //message 
-          await app.client.chat.scheduleMessage({
-            channel: process.env.CHANNEL_ID,
-            text: 'Tset Dome is away: Away for 1.00 day',
-            post_at
-          });
-        } else {
-          console.log('No Days off Today');
-        }
-      } catch(e) {
-        console.error(e);
-      }
-    }
-  }, timeInterval);
-}
 
 (async () => {
   const port = 3000
@@ -80,7 +49,8 @@ async function runTimeOffEvents(startTime) {
     await app.start(process.env.PORT || port);
     console.log(`⚡️ Time off bot is running on port ${port}!`);
     const startTime = dayjs().tz('America/Toronto');
-    runTimeOffEvents(startTime);
+    const timeInterval = 8.64e7;
+    runTimeOffEvents(app, startTime, timeInterval);
   } catch(e) {
     console.error(e);
   }
